feat: move scroll to active logic to a hook
